Serve WASM files using the correct content-type.
Live-server was using an older version of send, which depended on an
older version of mime, that served .wasm files using the content-type
`application/octet-stream` instead of `application/wasm`.

This breaks WebAssembly.instantiateStreaming, which expects the correct
mime type or the call will fail.

Upgrading the dependency serves WASM using the correct content-type, and
makes developing WASM applications in the browser much easier.